split test_basenode into themed modules
1620 lines and 61 tests covering creation, matching, filters, property types,
dataframes, related nodes and schema made failures hard to locate and the file
a merge-conflict magnet.

Split along those seams into create, match, properties, dataframes and related
modules. Coupling was low: only PracticeNode and SampleEnum were used by more
than one, and they now live in models_basenode.py - primary labels must be
unique across the suite, so a shared model has to be defined once.

All 421 test names and outcomes are identical to before the split.
